Use BigDecimal instead of double for prices and quantities

+++ src/main/java/com/alex/trading_engine/controller/dto/OrderBookResponse.java
@@ -1,9 +1,10 @@
 package com.alex.trading_engine.controller.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /** Response body for GET /orderbook: top bid and ask levels with price and total quantity. */
 public record OrderBookResponse(List<PriceLevel> bids, List<PriceLevel> asks) {
 
-    public record PriceLevel(double price, double totalQuantity) {}
+    public record PriceLevel(BigDecimal price, BigDecimal totalQuantity) {}
 }
