+++ src/main/java/com/alex/trading_engine/controller/dto/TradeResponse.java
@@ -2,6 +2,7 @@ package com.alex.trading_engine.controller.dto;
 
 import com.alex.trading_engine.model.Trade;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 /** Response body for GET /trades. */
@@ -9,8 +10,8 @@ public record TradeResponse(
         String buyerOrderId,
         String sellerOrderId,
         String symbol,
-        double price,
-        double quantity,
+        BigDecimal price,
+        BigDecimal quantity,
         Instant timestamp
 ) {
     public static TradeResponse from(Trade trade) {
